listen_jeenode/requirements.txt
docopt==0.6.2
hiredis==0.2.0
pyserial==3.0.1
redis==2.10.5
requests==2.9.1
setproctitle==1.1.9
